+++ src/common/models/raster-job.ts
@@ -1,5 +1,0 @@
-export enum RasterIngestionJobType {
-  NEW = 'Ingestion_New',
-  UPDATE = 'Ingestion_Update',
-  SWAP_UPDATE = 'Ingestion_Swap_Update',
-}