+++ src/discrete-layer/components/job-manager/cell-renderer/job-details/job-details.raster-job-data.tsx
@@ -1,6 +1,6 @@
 import { ICellRendererParams } from 'ag-grid-community';
 import { get } from 'lodash';
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { useIntl } from 'react-intl';
 import { Box } from '@map-colonies/react-components';
 import {
@@ -13,15 +13,20 @@ import {
 import { AutoDirectionBox } from '../../../../../common/components/auto-direction-box/auto-direction-box.component';
 import { DETAILS_ROW_ID_SUFFIX } from '../../../../../common/components/grid';
 import { Hyperlink } from '../../../../../common/components/hyperlink/hyperlink';
+import { useEnums } from '../../../../../common/hooks/useEnum.hook';
 import { Domain } from '../../../../../common/models/domain';
-import { RasterErrorsSummary } from '../../../../../common/models/job-errors-summary.raster';
-import { RasterIngestionJobType } from '../../../../../common/models/raster-job';
+import {
+  APPROVAL_REQUIRED_ERRORS,
+  RasterErrorsCountKey,
+  RasterErrorsSummary,
+} from '../../../../../common/models/job-errors-summary.raster';
 import { JobModelType, Status, TaskModelType, useStore } from '../../../../models';
 import useZoomLevelsTable from '../../../export-layer/hooks/useZoomLevelsTable';
 import {
   getRasterErrorCount,
-  JobErrorsSummary,
-} from '../../../job-errors-summary/job-errors-summary';
+  JobErrorsSummaryRasterJobData,
+} from './job-errors-summary.raster-job-data';
+import { getEnumWithRealValues } from '../../../layer-details/utils';
 
 import './info-area.css';
 import './job-details.raster-job-data.css';
@@ -32,25 +37,34 @@ const MAX_ERRORS_SHOWN = 99;
 
 export const JobDetailsRasterJobData: React.FC<JobDetailsRasterJobDataProps> = ({ data }) => {
   const jobData = data as JobModelType;
-  const store = useStore();
   const intl = useIntl();
   const theme = useTheme();
+  const store = useStore();
+  const ENUMS = useEnums();
   const ZOOM_LEVELS_TABLE = useZoomLevelsTable();
   const [task, setTask] = useState<TaskModelType>();
   const [errorsCount, setErrorsCount] = useState(0);
   const [zoomLevel, setZoomLevel] = useState<number | undefined>(undefined);
   const [isLoading, setIsLoading] = useState(true);
 
+  const rasterIngestionJobTypeRealValues = useMemo(() => {
+    return Object.values(getEnumWithRealValues(ENUMS, 'RasterIngestionJobType'));
+  }, []);
+
   const isRasterJob =
     jobData.domain === Domain.RASTER &&
     jobData.type &&
-    Object.values(RasterIngestionJobType).includes(jobData.type as RasterIngestionJobType);
+    rasterIngestionJobTypeRealValues.includes(jobData.type as string);
 
   const calculateErrorsCount = (errors: RasterErrorsSummary): number => {
     let count = 0;
     Object.keys(errors.errorsCount).forEach((key) => {
       const errorCount = getRasterErrorCount(errors, key);
-      if (errorCount.exceeded === true || typeof errorCount.exceeded === 'undefined') {
+      if (
+        APPROVAL_REQUIRED_ERRORS.includes(key as RasterErrorsCountKey) ||
+        errorCount.exceeded === true ||
+        typeof errorCount.exceeded === 'undefined'
+      ) {
         count += errorCount.count ?? 0;
       }
     });
@@ -150,7 +164,11 @@ export const JobDetailsRasterJobData: React.FC<JobDetailsRasterJobDataProps> = (
               <Tooltip
                 content={
                   <Box>
-                    {JobErrorsSummary(theme, task?.parameters?.errorsSummary, 'reportItem')}
+                    {JobErrorsSummaryRasterJobData(
+                      theme,
+                      task?.parameters?.errorsSummary,
+                      'reportItem'
+                    )}
                   </Box>
                 }
               >
@@ -180,7 +198,7 @@ export const JobDetailsRasterJobData: React.FC<JobDetailsRasterJobDataProps> = (
             <Tooltip
               content={
                 <Box>
-                  {JobErrorsSummary(
+                  {JobErrorsSummaryRasterJobData(
                     theme,
                     task?.parameters?.errorsSummary,
                     'reportItem',